Implement SSE as a response
