+++ src/response/sse.rs
@@ -1,0 +1,346 @@
+//! Server-Sent Events (SSE) responses.
+//!
+//! # Example
+//!
+//! ```
+//! use axum::prelude::*;
+//! use axum::response::sse::{sse, Event, KeepAlive, Sse};
+//! use std::{time::Duration, convert::Infallible};
+//! use tokio_stream::StreamExt as _ ;
+//! use futures::stream::{self, Stream};
+//!
+//! let app = route("/sse", get(sse_handler));
+//!
+//! async fn sse_handler() -> Sse<impl Stream<Item = Result<Event, Infallible>>> {
+//!     // A `Stream` that repeats an event every second
+//!     let stream = stream::repeat_with(|| Event::default().data("hi!"))
+//!         .map(Ok)
+//!         .throttle(Duration::from_secs(1));
+//!
+//!     sse(stream).keep_alive(KeepAlive::default())
+//! }
+//! # async {
+//! # hyper::Server::bind(&"".parse().unwrap()).serve(app.into_make_service()).await.unwrap();
+//! # };
+//! ```
+
+use crate::{body::BoxStdError, response::IntoResponse};
+use futures_util::stream::{Stream, StreamExt as _, TryStream, TryStreamExt as _};
+use http::Response;
+use hyper::Body;
+use pin_project_lite::pin_project;
+use serde::Serialize;
+use std::{
+    borrow::Cow,
+    fmt,
+    fmt::Write,
+    future::Future,
+    pin::Pin,
+    task::{Context, Poll},
+    time::Duration,
+};
+use tokio::time::Sleep;
+use tower::BoxError;
+
+/// Create a new [`Sse`] response that will respond with the given stream of
+/// [`Event`]s.
+///
+/// See the [module docs](self) for more details.
+pub fn sse<S>(stream: S) -> Sse<S>
+where
+    S: TryStream<Ok = Event> + Send + 'static,
+    S::Error: Into<BoxError>,
+{
+    Sse {
+        stream,
+        keep_alive: None,
+    }
+}
+
+/// An SSE response.
+#[derive(Clone)]
+pub struct Sse<S> {
+    stream: S,
+    keep_alive: Option<KeepAlive>,
+}
+
+impl<S> Sse<S> {
+    /// Configure the interval between keep-alive messages.
+    ///
+    /// Defaults to no keep-alive messages.
+    pub fn keep_alive(mut self, keep_alive: KeepAlive) -> Self {
+        self.keep_alive = Some(keep_alive);
+        self
+    }
+}
+
+impl<S> fmt::Debug for Sse<S> {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        f.debug_struct("Sse")
+            .field("stream", &format_args!("{}", std::any::type_name::<S>()))
+            .field("keep_alive", &self.keep_alive)
+            .finish()
+    }
+}
+
+impl<S, E> IntoResponse for Sse<S>
+where
+    S: Stream<Item = Result<Event, E>> + Send + 'static,
+    E: Into<BoxError>,
+{
+    fn into_response(self) -> Response<Body> {
+        let stream = if let Some(keep_alive) = self.keep_alive {
+            KeepAliveStream {
+                event_stream: self.stream,
+                comment_text: keep_alive.comment_text,
+                max_interval: keep_alive.max_interval,
+                alive_timer: tokio::time::sleep(keep_alive.max_interval),
+            }
+            .left_stream::<S>()
+        } else {
+            self.stream.right_stream::<KeepAliveStream<S>>()
+        };
+
+        let stream = stream
+            .map_ok(|event| event.to_string())
+            .map_err(|err| BoxStdError(err.into()));
+
+        Response::builder()
+            .header(http::header::CONTENT_TYPE, "text/event-stream")
+            .header(http::header::CACHE_CONTROL, "no-cache")
+            .body(Body::wrap_stream(stream))
+            .unwrap()
+    }
+}
+
+/// Server-sent event
+#[derive(Default, Debug)]
+pub struct Event {
+    name: Option<String>,
+    id: Option<String>,
+    data: Option<DataType>,
+    event: Option<String>,
+    comment: Option<String>,
+    retry: Option<Duration>,
+}
+
+// Server-sent event data type
+#[derive(Debug)]
+enum DataType {
+    Text(String),
+    Json(String),
+}
+
+impl Event {
+    /// Set Server-sent event data
+    /// data field(s) ("data:<content>")
+    pub fn data<T>(mut self, data: T) -> Event
+    where
+        T: Into<String>,
+    {
+        self.data = Some(DataType::Text(data.into()));
+        self
+    }
+
+    /// Set Server-sent event data
+    /// data field(s) ("data:<content>")
+    pub fn json_data<T>(mut self, data: T) -> Result<Event, serde_json::Error>
+    where
+        T: Serialize,
+    {
+        self.data = Some(DataType::Json(serde_json::to_string(&data)?));
+        Ok(self)
+    }
+
+    /// Set Server-sent event comment
+    /// Comment field (":<comment-text>")
+    pub fn comment<T>(mut self, comment: T) -> Event
+    where
+        T: Into<String>,
+    {
+        self.comment = Some(comment.into());
+        self
+    }
+
+    /// Set Server-sent event event
+    /// Event name field ("event:<event-name>")
+    pub fn event<T>(mut self, event: T) -> Event
+    where
+        T: Into<String>,
+    {
+        self.event = Some(event.into());
+        self
+    }
+
+    /// Set Server-sent event retry
+    /// Retry timeout field ("retry:<timeout>")
+    pub fn retry(mut self, duration: Duration) -> Event {
+        self.retry = Some(duration);
+        self
+    }
+
+    /// Set Server-sent event id
+    /// Identifier field ("id:<identifier>")
+    pub fn id<T>(mut self, id: T) -> Event
+    where
+        T: Into<String>,
+    {
+        self.id = Some(id.into());
+        self
+    }
+}
+
+impl fmt::Display for Event {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        if let Some(comment) = &self.comment {
+            ":".fmt(f)?;
+            comment.fmt(f)?;
+            f.write_char('\n')?;
+        }
+
+        if let Some(event) = &self.event {
+            "event:".fmt(f)?;
+            event.fmt(f)?;
+            f.write_char('\n')?;
+        }
+
+        match &self.data {
+            Some(DataType::Text(data)) => {
+                for line in data.split('\n') {
+                    "data:".fmt(f)?;
+                    line.fmt(f)?;
+                    f.write_char('\n')?;
+                }
+            }
+            Some(DataType::Json(data)) => {
+                "data:".fmt(f)?;
+                data.fmt(f)?;
+                f.write_char('\n')?;
+            }
+            None => {}
+        }
+
+        if let Some(id) = &self.id {
+            "id:".fmt(f)?;
+            id.fmt(f)?;
+            f.write_char('\n')?;
+        }
+
+        if let Some(duration) = &self.retry {
+            "retry:".fmt(f)?;
+
+            let secs = duration.as_secs();
+            let millis = duration.subsec_millis();
+
+            if secs > 0 {
+                // format seconds
+                secs.fmt(f)?;
+
+                // pad milliseconds
+                if millis < 10 {
+                    f.write_str("00")?;
+                } else if millis < 100 {
+                    f.write_char('0')?;
+                }
+            }
+
+            // format milliseconds
+            millis.fmt(f)?;
+
+            f.write_char('\n')?;
+        }
+
+        f.write_char('\n')?;
+
+        Ok(())
+    }
+}
+
+/// Configure the interval between keep-alive messages, the content
+/// of each message, and the associated stream.
+#[derive(Debug, Clone)]
+pub struct KeepAlive {
+    comment_text: Cow<'static, str>,
+    max_interval: Duration,
+}
+
+impl KeepAlive {
+    /// Create a new `KeepAlive`.
+    pub fn new() -> Self {
+        Self {
+            comment_text: Cow::Borrowed(""),
+            max_interval: Duration::from_secs(15),
+        }
+    }
+
+    /// Customize the interval between keep-alive messages.
+    ///
+    /// Default is 15 seconds.
+    pub fn interval(mut self, time: Duration) -> Self {
+        self.max_interval = time;
+        self
+    }
+
+    /// Customize the text of the keep-alive message.
+    ///
+    /// Default is an empty comment.
+    pub fn text<I>(mut self, text: I) -> Self
+    where
+        I: Into<Cow<'static, str>>,
+    {
+        self.comment_text = text.into();
+        self
+    }
+}
+
+impl Default for KeepAlive {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+pin_project! {
+    struct KeepAliveStream<S> {
+        #[pin]
+        event_stream: S,
+        comment_text: Cow<'static, str>,
+        max_interval: Duration,
+        #[pin]
+        alive_timer: Sleep,
+    }
+}
+
+impl<S> Stream for KeepAliveStream<S>
+where
+    S: TryStream<Ok = Event>,
+{
+    type Item = Result<Event, S::Error>;
+
+    fn poll_next(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Option<Self::Item>> {
+        let mut this = self.project();
+
+        match this.event_stream.try_poll_next(cx) {
+            Poll::Pending => match this.alive_timer.as_mut().poll(cx) {
+                Poll::Pending => Poll::Pending,
+                Poll::Ready(_) => {
+                    // restart timer
+                    this.alive_timer
+                        .reset(tokio::time::Instant::now() + *this.max_interval);
+
+                    let comment_str = this.comment_text.clone();
+                    let event = Event::default().comment(comment_str);
+                    Poll::Ready(Some(Ok(event)))
+                }
+            },
+            Poll::Ready(Some(Ok(event))) => {
+                // restart timer
+                this.alive_timer
+                    .reset(tokio::time::Instant::now() + *this.max_interval);
+
+                Poll::Ready(Some(Ok(event)))
+            }
+            Poll::Ready(None) => Poll::Ready(None),
+            Poll::Ready(Some(Err(error))) => Poll::Ready(Some(Err(error))),
+        }
+    }
+}
